flask/requirements.txt
flask
Flask-SQLAlchemy
psycopg2
psycopg2-binary
alembic
openai